+++ src/patterns/relay.ts
@@ -99,7 +99,7 @@
  * @module
  */
 
-import { immutable } from "../values/structures.js";
+import { immutable } from "../types/values.js";
 import { isDefined, isFunction } from "../index.js";